Random Commit.
git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@202 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ ENIGMAsystem/SHELL/Universal_System/mathnc.h
@@ -79,7 +79,7 @@ extern double random(double n);
 int random_set_seed(int seed);
 int random_get_seed();
 int randomize();
-int random_integer(double x);
+int random_integer(int x);
 int random32();
 double mtrandom();
 int mtrandom_seed(int x);
